music_analyzer: Fixed GUI "Not Responding" when clicking 'Analyze'
The 'Analyze' functions can take a relatively long
time to execute and while it is running Windows is
periodically checking if the GUI is alive. Because
'Analyze' takes a while to complete the process is
not acknowleding any Winows is alive signals and
Windows marks the program as 'Not Responding' and
will prompt the user if they wish to terminate it
if they spam click on the GUI. This commit aims to
fix that.

'Analyze' is run as a seperate daemon thread from
the main program so when Windows checks for life
for this program the main thread is able to
respond and Windows wont mark the program as 'Not
Responding'.

The overall execution time is still the same as it
was before. The performance of 'Analyze' is not
acceptable and needs to be much faster, in the
order of milliseconds NOT seconds, but that can be
a todo later item.

	modified:   music_analyzer/app.py